Rename to more generic "terraform"
I know that this might fit the name of "terraform-basic", but there's no other "terraform-advanced" to need such differentiation. If/when such a
template comes to exist, this can and SHOULD be named more specifically. However, in this context, it doesn't need to be
"terraform-basic" and can instead be the shorter and more succinct "terraform".