Add comments and refine devcontainer.json
This removes a LOT of the options that were there before. I have bumped the version to 2.0.0 since this is a breaking change.
I don't know this is a good idea? Maybe it is? Ideally I think (OPINION) that the devcontainer templates should be as lean as possible in terms of config maybe with json comments on how to customize it.